+++ src/Exception/InvalidValueRangeException.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\Colour\Exception;
-
-final class InvalidValueRangeException extends DomainException
-{
-}