+++ tests/BlueTest.php
@@ -6,7 +6,6 @@ namespace Tests\Innmind\Colour;
 use Innmind\Colour\{
     Blue,
     Intensity,
-    Exception\InvalidValueRangeException,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -60,7 +59,7 @@ class BlueTest extends TestCase
 
     public function testThrowWhenValueIsTooLow()
     {
-        $this->expectException(InvalidValueRangeException::class);
+        $this->expectException(\OutOfBoundsException::class);
         $this->expectExceptionMessage('-42');
 
         Blue::of(-42)->unwrap();
@@ -68,7 +67,7 @@ class BlueTest extends TestCase
 
     public function testThrowWhenValueIsTooHigh()
     {
-        $this->expectException(InvalidValueRangeException::class);
+        $this->expectException(\OutOfBoundsException::class);
         $this->expectExceptionMessage('512');
 
         Blue::of(512)->unwrap();