+++ tests/RGBATest.php
@@ -11,7 +11,6 @@ use Innmind\Colour\{
     RGBA,
     HSLA,
     CMYKA,
-    Exception\DomainException,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -310,7 +309,7 @@ class RGBATest extends TestCase
 
     public function testThrowWhenInvalidRGBFunctionWithPoints()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('rgb(10, 20%, 30)');
 
         RGBA::of('rgb(10, 20%, 30)');
@@ -329,7 +328,7 @@ class RGBATest extends TestCase
 
     public function testThrowWhenInvalidRGBFunctionWithPercents()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('rgb(10, 20%, 30)');
 
         RGBA::of('rgb(10, 20%, 30)');
@@ -367,7 +366,7 @@ class RGBATest extends TestCase
 
     public function testThrowWhenInvalidRGBAFunctionWithPoints()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('rgba(10, 20%, 30, 2.0)');
 
         RGBA::of('rgba(10, 20%, 30, 2.0)');
@@ -386,7 +385,7 @@ class RGBATest extends TestCase
 
     public function testThrowWhenInvalidRGBAFunctionWithPercents()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('rgba(10, 20%, 30, 1)');
 
         RGBA::of('rgba(10, 20%, 30, 1)');